backend/requirements.txt
# ULTRA-SIMPLE REQUIREMENTS - Compatible with Python 3.14+
setuptools>=69.0.2
fastapi>=0.111.0
uvicorn>=0.24.0
h11>=0.8
httptools>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

email-validator==2.1.0
python-multipart==0.0.9

# Use requests instead of supabase library (no httpx conflicts!)
requests==2.31.0

# PDF processing and generation
PyPDF2==3.0.1
reportlab==4.0.7

# Excel export with formatting
openpyxl==3.1.2

# Image processing (for PDFs with images) - Updated for Python 3.14+ compatibility
pillow>=10.1.0

# Payment processing & gateway
razorpay>=2.0.0

# Supabase client
supabase>=2.5.0

# AI Services for extraction
openai==1.30.1
google-generativeai>=0.8.0
google-cloud-vision>=3.4.0

# Database ORM
sqlalchemy>=2.0.44
psycopg2-binary>=2.9.9


# Authentication & Security
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Rate Limiting
slowapi==0.1.9

# Error Monitoring & Tracking
sentry-sdk[fastapi]==1.40.6

# Settings management
# (pydantic 1.10 has BaseSettings built-in, no separate pydantic-settings needed)
